changing testing structure
